Introduce Safe write and clear state functions
These shall allow implementing consistent writes of state that can be updated
 by multiple writers.

The current change set modifies the OvsNwOperState writes and clears to
the safe version.

This is part of initial commits that enables cleaninup delete API of the
core to just take keys, since now oper (or config) state can be guranteed
to exist until all writers are done with it.